Update search by name
